+++ ICSSoft.STORMNET.Business/DataServiceProvider.cs
@@ -4,8 +4,6 @@
     using System.Configuration;
     using System.Security.Cryptography;
     using System.Text;
-    using System.Web;
-    using System.Web.Configuration;
     using ICSSoft.Services;
 
     using Unity;
@@ -73,13 +71,11 @@
                             if (dataService.CustomizationString == null)
                             {
                                 // Пробуем получить строку соединения в web-стиле.
-                                string connectionStringName =
-                                    WebConfigurationManager.AppSettings["DefaultConnectionStringName"];
+                                string connectionStringName = ConfigurationManager.AppSettings["DefaultConnectionStringName"];
                                 string connectionString = null;
                                 if (!string.IsNullOrEmpty(connectionStringName))
                                 {
-                                    ConnectionStringSettings connString =
-                                        WebConfigurationManager.ConnectionStrings[connectionStringName];
+                                    ConnectionStringSettings connString = ConfigurationManager.ConnectionStrings[connectionStringName];
                                     if (connString != null)
                                     {
                                         connectionString = connString.ConnectionString;
@@ -182,23 +178,13 @@
         }
 
         /// <summary>
-        /// Private field for <see cref="IsWebApp"/> property.
-        /// </summary>
-        private static bool? _isWebApp;
-
-        /// <summary>
         /// Determinate is a web application. Will be check <see cref="HttpContext.Current"/> is not null for web applications.
         /// </summary>
         public static bool IsWebApp
         {
             get
             {
-                if (_isWebApp == null)
-                {
-                    _isWebApp = HttpContext.Current != null;
-                }
-
-                return _isWebApp.Value;
+                return false;
             }
         }